almost ready to submit
Camera ready copy is due in a week.  I have addressed the reviewers'
concerns by expanding the justification for the completeness of the
model and showing how it explains collateral, compartmented, and
international accreditations.  I showed how the proposed solution
addresses the cost of certification by giving the developer control of
the schedule, and of accreditation by predicting the behaviour of
accreditors.  Proofread it in the morning and if it looks OK, then ship
it.